lections/portfolio-api/requirements.txt
django==4.1.5
psycopg2-binary==2.9.5
python-dotenv==0.21.0
django-environ==0.9.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.4
pillow==9.4.0
whitenoise==6.3.0
